Reset Archive in MOSuiteStrategy when using SELECTED_JUNIT because we do not want the targets to be added twice. Also, remove hours from the name of CoveredGoalsOverTime.csv.
When using MOSuiteStrategy with prior test seeding, we end up with two sets of targets because they are added by the carving manager and the MOSuiteStrategy. To prevent this duplication, we reset the archive in MOSuiteStrategy.

Previously I added hours to the name of the csv recording covered targets per iteration. Now removing it because it is no longer useful.