Fix "Exception" target duplication in the CoverageArchive.
Previously when adding a new target of type exception, the algorithm checked for its presence in the uncovered archive. However, exception targets become covered by default when they are discovered, so they are instantly added to the covered archive. But the algorithm did not check for its presence in the covered archive, leading to duplication of "exception" type targets.