Change CoveredGoalsOverTime.csv to record timestamps rather than iteration numbers. Also, move writing to the csv file operation at the end of the search.
Previously, the CoveredGoalsOverTime.csv file kept a record of iteration index and the goals achieved in this iteration. This was recorded whenever more goals would be achieved. However, this approach is less precise because we cannot guarantee that each iteration took the same time to complete, therefore, it is better to use a timestamp.

Additionally, I moved writing to the csv file to be at the end of the search instead of during it. The search now adds all values to the array list and later writes it to the file thereby spending less search time on non-search-related activity.